Use number of associations instead of positive balances

+++ hedera-node/hedera-config/src/main/java/com/hedera/node/config/data/TokensConfig.java
@@ -31,7 +31,7 @@ public record TokensConfig(
         @ConfigProperty(defaultValue = "100") @NetworkProperty int maxTokenNameUtf8Bytes,
         @ConfigProperty(defaultValue = "10") @NetworkProperty int maxCustomFeesAllowed,
         @ConfigProperty(defaultValue = "2") @NetworkProperty int maxCustomFeeDepth,
-        @ConfigProperty(defaultValue = "1000") @NetworkProperty long maxRelsPerInfoQuery,
+        @ConfigProperty(defaultValue = "1000") @NetworkProperty int maxRelsPerInfoQuery,
         @ConfigProperty(value = "reject.enabled", defaultValue = "true") @NetworkProperty boolean tokenRejectEnabled,
         @ConfigProperty(value = "nfts.areEnabled", defaultValue = "true") @NetworkProperty boolean nftsAreEnabled,
         @ConfigProperty(value = "nfts.maxMetadataBytes", defaultValue = "100") @NetworkProperty
